Formatting in Templates and README
I have made a correction to the Templates by removing the accidental reference to April 2025 notes. I originally built the template off of those notes and didn't notice this. Good thing I caught this in the drafts!

Additionally, I updated the README for both more clarity and to add the project workflow.